+++ ChangeLog.md
@@ -5,6 +5,7 @@ Incoming in 0.9.9
 -----------------
 * Features
     * Add `MatchGroup.equals()`, that allows to compare a potentially unmatched group with a string. 
+    * Add `pattern()->match()->group()->filter()` method. #22
 * Enhancements
     * [Prepared patterns] correctly handle whitespace when `x` (`PCRE_EXTENDED`) flag is used. #40
 * SafeRegex